requirements.txt
crewai>=0.16.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.5.0
